Move semantic token handle to its own class
